+++ src/pages/image-sitemap.xml.ts
@@ -1,6 +1,7 @@
 import type { APIRoute } from 'astro';
 import { getCollection } from 'astro:content';
 import { site } from '../config/site';
+import seoRelease from '../config/seo-release.json';
 
 export const prerender = true;
 
@@ -11,12 +12,12 @@ export const GET: APIRoute = async () => {
   const entries = [];
   for (const locale of site.locales) {
     for (const { data: product } of products) {
-      entries.push({ page: `/${locale}/product/${product.slug}/`, image: product.image, title: product.translations[locale].name });
+      entries.push({ page: `/${locale}/product/${product.slug}/`, image: product.image, lastmod: null });
     }
     for (const { data: post } of posts) {
-      if (post.image) entries.push({ page: `/${locale}/blog/${post.translations[locale].slug}/`, image: post.image, title: post.translations[locale].title });
+      if (post.image) entries.push({ page: `/${locale}/blog/${post.translations[locale].slug}/`, image: post.image, lastmod: seoRelease.lastSignificantUpdate });
     }
   }
-  const xml = `<?xml version="1.0" encoding="UTF-8"?>\n<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1">\n${entries.map(({ page, image, title }) => `  <url><loc>${site.origin}${escapeXml(page)}</loc><image:image><image:loc>${site.origin}${escapeXml(image)}</image:loc><image:title>${escapeXml(title)}</image:title></image:image></url>`).join('\n')}\n</urlset>\n`;
+  const xml = `<?xml version="1.0" encoding="UTF-8"?>\n<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1">\n${entries.map(({ page, image, lastmod }) => `  <url><loc>${site.origin}${escapeXml(page)}</loc>${lastmod ? `<lastmod>${lastmod}</lastmod>` : ''}<image:image><image:loc>${site.origin}${escapeXml(image)}</image:loc></image:image></url>`).join('\n')}\n</urlset>\n`;
   return new Response(xml, { headers: { 'Content-Type': 'application/xml; charset=utf-8' } });
 };